% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-summary.R
\name{as.maars}
\alias{as.maars}
\alias{as.maars.lm}
\title{Convert an object to an object that can be handled by the "maars" package}
\usage{
as.maars(x, ...)

\method{as.maars}{lm}(x, ...)
}
\arguments{
\item{x}{Object to be converted. See Methods section below for details on
formatting of each input type.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Several methods are provided to convert common objects
(such as "lm") into "maars" objects,
which can be used with the various functions in the package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.maars}: The input object \code{x} must be of class
\code{\link{stats::lm}}. THe function returns an object of class
("maars_lm, "lm").
}}

