% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical-var.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Draw bootstrap samples from the data}
\usage{
bootstrap_samples(data, B = 100, m = NULL)
}
\arguments{
\item{data}{Tibble or data frame containing the dataset to be sampled from}

\item{B}{Bootstrap repetitions or number of bootstrap samples to be drawn}

\item{m}{Number of observations to be sampled with replacement from the dataset
for each bootstrap repetition}
}
\value{
A tibble of the bootstrap samples
}
\description{
Draw bootstrap samples from the data.
}
\examples{
\dontrun{
# Draw B=100 bootstrap samples of size m=5
set.seed(162632)
n <- 100
boot <- bootstrap_samples(data.frame(y = rnorm(n,0,1), x = rnorm(n,0,1)), B = 100, m = 5)

# Display the output
print(boot)
}
}
