% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-summary.R
\name{check_fn_args_summary}
\alias{check_fn_args_summary}
\title{Assertion Checks for individual \code{\link{get_summary}},
\code{\link{get_assumptions}}, \code{\link{get_summary}} function inputs}
\usage{
check_fn_args_summary(
  mod_fit,
  sand,
  boot_emp,
  boot_res,
  boot_mul,
  well_specified
)
}
\arguments{
\item{mod_fit}{(maars_lm, lm) A fitted OLS \code{maars_lm, lm} class object}

\item{sand}{(logical) : \code{TRUE} if sandwich estimator output is required,
\code{FALSE} to exclude this output from the request}

\item{boot_emp}{(logical) : \code{TRUE} if empirical bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{boot_res}{(logical) : \code{TRUE} if residual bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{boot_mul}{(logical) : \code{TRUE} if multiplier bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{well_specified}{(logical) : \code{TRUE} if lm standard errors
(well specified) output is required, \code{FALSE} to exclude this output
from the request}
}
\value{
: A character vector of  \code{TRUE} if assertions pass, or an
error if there is an assertion failure.
}
\description{
\code{check_fn_args_comp_var_boot_ind} is used to assess whether the arguments
are correctly specified in \code{list} format and returns an error message if
they do not match the correct specification
}
\examples{
\dontrun{
set.seed(1243434)

# generate data
n <- 1e3
X_1 <- stats::rnorm(n, 0, 1)
X_2 <- stats::rnorm(n, 10, 20)
eps <- stats::rnorm(n, 0, 1)

# OLS data and model
y <- 2 + X_1 * 1 + X_2 * 5 + eps
lm_fit <- stats::lm(y ~ X_1 + X_2)

# Create a maars_lm model with empirical bootstrap, sandwich, and
# well-specified variance estimates
mss_var1 <- mss_var(mod_fit = lm_fit, boot_emp = list(B = 20, m = 200),
                    boot_res = NULL,
                    boot_mul = NULL)

# Let the user request only the empirical bootstrap variance estimates
# This function should return "boot_emp" since the user only requested it
# and it is available in the fitted maars_lm model
check_fn_args_summary(mod_fit = mss_var1, sand = FALSE,
                      boot_emp = TRUE, boot_res = FALSE, boot_mul = FALSE,
                      well_specified = FALSE)
}
}
\keyword{internal}
