% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical.R
\name{comp_boot_emp_samples}
\alias{comp_boot_emp_samples}
\title{Draw bootstrap sample from the data}
\usage{
comp_boot_emp_samples(data, B = 1, m = NULL, replace = TRUE)
}
\arguments{
\item{data}{A tibble or data frame containing the dataset to be
sampled from.}

\item{B}{Bootstrap repetitions or number of bootstrap samples to
be drawn.}

\item{m}{Number of observations to be sampled with replacement from
the original dataset for each bootstrap repetition.}

\item{replace}{TODO: ADD}
}
\value{
A tibble containing the number of bootstrap iteration, the
bootstrap samples (\code{data}),
and the indices of the observations (linked to the original sample) in the
sample (\code{indices}).
}
\description{
\code{comp_boot_emp_sample} draws one bootstrap samples
(i.e. sampling observations with replacement) from the input
\code{data}. It returns a list containing the indices of the
\code{m} observations that have been sampled (\code{indices}) and
the bootstrapped dataset (\code{data}).
}
\examples{
\dontrun{
# Draw B=100 bootstrap samples of size m=5
set.seed(162632)
n <- 100
boot <- comp_boot_emp_samples(
  data.frame(
    y = stats::rnorm(n, 0, 1),
    x = stats::rnorm(n, 0, 1)
  ),
  B = 100,
  m = 5
)

# Display the output
print(boot)
}
}
\keyword{internal}
