% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-boot-multiplier.R
\name{comp_boot_mul_wgt}
\alias{comp_boot_mul_wgt}
\title{Generate Different Types of multiplier bootstrap weights}
\usage{
comp_boot_mul_wgt(n, weights_type)
}
\arguments{
\item{n}{The number of random multiplier bootstrap weights to generate.}

\item{weights_type}{The type of multiplier bootstrap weights to generate.
Based on the \code{weighttype} option in the \code{Stata boottest} package,
this can only take the following five pre-specified values
\code{"rademacher", "mammen", "webb", "std_gaussian", "gamma"}.
A brief description of each type is as follows. The \code{"rademacher"}
weights are sampled from the two point Rademacher distribution which
takes values \eqn{\{1, 1\}}, with equal probability. The \code{"mammen"}
weights are sampled from the two point Mammen distribution which has takes
values \eqn{\{\phi, 1 - \phi\}}, with probabilities
\eqn{\{\frac{\phi}{\sqrt{5}}, 1 - \frac{\phi}{\sqrt{5}}\}}, respectively.
The \code{"webb"} weights are sampled from the six point Webb distribution
which has takes values
\eqn{\{\pm \sqrt{\frac{3}{2}}, \pm \sqrt{\frac{1}{2}}, \pm 1 \}},
with equal probability. The \code{"std_gaussian"} weights are sampled from
the standard Gaussian (normal) distribution with mean = \eqn{0},
and variance = \eqn{1}. Finally, the \code{"gamma"} weights are sampled
from the Gamma distribution with shape parameter = \eqn{4}, and
scale parameter = \eqn{1}. For more details on these weight types see
\insertCite{@see @roodman2019fastwildinferencestataboottest;textual}{maars}.}
}
\value{
A numeric vector of n (sampled with replacement) random multiplier
bootstrap weights based on the specified multiplier weights type.
}
\description{
\code{comp_boot_mul_wgt} is a Helper function for
\code{\link{comp_boot_mul}} to generate different types
of multiplier bootstrap weights. This section is inspired by the
\code{weighttype} option in the Stata \code{boottest} package.
}
\examples{
\dontrun{
set.seed(824908)
# Number of multiplier weights to generate
n <- 1000

# Generate the different type of multiplier weights
rademacher_w <- comp_boot_mul_wgt(
  n = n,
  weights_type = "rademacher"
)
mammen_w <- comp_boot_mul_wgt(n = n, weights_type = "mammen")
webb_w <- comp_boot_mul_wgt(n = n, weights_type = "webb")
std_gaussian_w <- comp_boot_mul_wgt(
  n = n,
  weights_type = "std_gaussian"
)
gamma_w <- comp_boot_mul_wgt(n = n, weights_type = "gamma")
}
}
\references{
\insertAllCited
}
\keyword{internal}
