% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-boot-summary.R
\name{comp_bootstrap_summary}
\alias{comp_bootstrap_summary}
\title{Create an empirical or multiplier bootstrap summary table}
\usage{
comp_bootstrap_summary(mod_fit, boot_out, boot_type)
}
\arguments{
\item{mod_fit}{An lm (OLS) object}

\item{boot_out}{A tibble of the bootstrap calculations from either
\code{\link{comp_empirical_bootstrap}} or
\code{\link{comp_multiplier_bootstrap_var}}, respectively.}

\item{boot_type}{A character specifying the bootstrap type. Must be "emp"
for output from \code{\link{comp_empirical_bootstrap}} or "mult"
for output from \code{\link{comp_multiplier_bootstrap_var}}, respectively.}
}
\value{
A summary statistics tibble for the bootstrap input.
}
\description{
\code{comp_bootstrap_summary} creates an OLS bootstrap summary table for
empirical bootstrap output or multiplier bootstrap output as generated from
\code{\link{comp_empirical_bootstrap}} or
\code{\link{comp_multiplier_bootstrap_var}}, respectively.
}
\examples{
\dontrun{
# Simulate data from a linear model
set.seed(35542)
n <- 1e2
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)

# Fit the linear model using OLS (ordinary least squares)
mod_fit <- stats::lm(y ~ X)

# Run the multiplier bootstrap on the fitted (OLS) linear model
set.seed(162632)
coef_emp_boot <- comp_empirical_bootstrap(mod_fit = mod_fit, B = 1000)
out <- comp_bootstrap_summary(mod_fit, coef_emp_boot, 'emp')

# print output
print(out)
}
}
