% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical-var.R
\name{comp_cond_model}
\alias{comp_cond_model}
\title{Fit OLS or GLM on the data}
\usage{
comp_cond_model(mod_fit, data, weights = NULL)
}
\arguments{
\item{mod_fit}{An object of class \code{\link[stats]{lm}} or
\code{\link[stats]{glm}} to fit on the data. This object should contain
the formula, the data, and, in case of \code{\link[stats]{glm}}, the
glm family.}

\item{data}{A tibble or data frame containing the dataset on which the
model will be trained.}

\item{weights}{A character corresponding to the name of the weights feature
name in the data.}
}
\value{
A tibble containing the estimated coefficients (\code{term}) of
the regressors (\code{term}).
}
\description{
\code{comp_cond_model} fits an OLS or GLM on the input dataset.
}
\details{
The model specification obtained from \code{mod_fit}, of
class \code{\link[stats]{lm}} or \code{\link[stats]{glm}}, is
fitted on \code{data}. The user can choose to fit a weighted
regression using the argument \code{weights}.
}
\examples{
\dontrun{
set.seed(457468)
# Estimate ols from a bootstraped dataset
n <- 1e3
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)
lm_fit <- stats::lm(y ~ X)
boot <- comp_empirical_bootstrap_samples(model.frame(lm_fit))$data[[1]]
mod <- comp_cond_model(lm_fit, boot)

# Display the output
print(mod)

#' # Estimate OLS from a dataset with weights
n <- 1e3
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)
mod_fit <- lm(y ~ X)
df <- tibble::tibble(y = y, X = X, weights = 1:length(X))
mod <- comp_cond_model(mod_fit, df, "weights")

# Display the output
print(mod)
}
}
