% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical-var.R
\name{comp_empirical_bootstrap}
\alias{comp_empirical_bootstrap}
\title{A wrapper for the empirical bootstrap of a fitted OLS or GLM models}
\usage{
comp_empirical_bootstrap(mod_fit, B = 100, m = NULL)
}
\arguments{
\item{mod_fit}{An object of class "lm" or "glm" to fit on the data. This object
should contain the formula, the data, and, in case of "glm", the family}

\item{B}{Bootstrap repetitions or number of bootstrap samples to be drawn}

\item{m}{Number of observations to be sampled with replacement from the dataset
for each bootstrap repetition}
}
\value{
A tibble of the model's coefficients estimated on the bootstrapped data sets
}
\description{
The empirical bootstrap consists of fitting the chosen statistical model
onto B bootstrap versions of size m of the data set
}
\examples{
\dontrun{
# Obtain estimates of the coefficients on bootstrapped versions of the data set
set.seed(35542)
n <- 1e3
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)
lm_fit <- stats::lm(y ~ X)
out <- comp_empirical_bootstrap(lm_fit, B = 100, m = 300)

print(out)
}
}
