% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical-var.R
\name{comp_empirical_bootstrap_samples}
\alias{comp_empirical_bootstrap_samples}
\title{Draw bootstrap samples from the data.}
\usage{
comp_empirical_bootstrap_samples(data, B = 100, m = NULL)
}
\arguments{
\item{data}{Tibble or data frame containing the dataset to be sampled from.}

\item{B}{Bootstrap repetitions or number of bootstrap samples to be drawn.}

\item{m}{Number of observations to be sampled with replacement from the original
data set for each bootstrap repetition.}
}
\value{
A tibble containing the bootstrap samples (\code{data})
and their corresponding number (\code{b}).
}
\description{
\code{comp_empirical_bootstrap_samples} draws bootstrap samples from the data.
}
\details{
More specifically, \code{B} bootstraped data sets of size \code{m} are drawn
from \code{data}.
}
\examples{
\dontrun{
# Draw B=100 bootstrap samples of size m=5
set.seed(162632)
n <- 100
boot <- comp_empirical_bootstrap_samples (data.frame(y = rnorm(n, 0, 1), x = rnorm(n, 0, 1)), B = 100, m = 5)

# Display the output
print(boot)
}
}
