% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{comp_grid_centers}
\alias{comp_grid_centers}
\title{Create grid of centers for reweighting the distribution of the regressor}
\usage{
comp_grid_centers(x, grid_method, n_grid)
}
\arguments{
\item{x}{Vector of values of the regressor}

\item{grid_method}{Method to construct the grid ("regular" or "quantiles")}

\item{n_grid}{Number of centers present in the grid}
}
\value{
A vector of values corresponding to the centers for the reweighting
}
\description{
Create grid of centers for reweighting the distribution of the regressor.
The function creates either a grid of evenly spaced values between the
minimum and the maximum of the regressor's values or based on the quantiles.
If based on the quantiles, then the values of the grid consists
of the quantiles corresponding
to seq(0.1,0.9,length=n_grid).
}
\examples{
\dontrun{
# Construct a grid of centers for the reweighting of the regressor
set.seed(162632)
n <- 100
x <- rnorm(n, 0, 1)
centers <- comp_grid_centers(x, "quantiles", 9)

# Display the output
print(centers)
}
}
