% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-var.R
\name{comp_mms_var}
\alias{comp_mms_var}
\title{Generates list containing several estimates of the variance}
\usage{
comp_mms_var(
  mod_fit,
  boot_emp = NULL,
  boot_sub = NULL,
  boot_res = NULL,
  boot_mul = NULL
)
}
\arguments{
\item{mod_fit}{An lm (OLS) object}

\item{boot_emp}{(list) In the case of empirical bootstrap the expected input
is of the form #'   \code{list(B = 10, m = 100)}. Here the named
element \code{m} is optional e.g. \code{list(B = 10)} is valid, or passed
in as an explicit \code{NULL} e.g. \code{list(B = 10, m = NULL)}.
Note that technically \code{B, m} should both be positive integers,
but this assertion checking is handled explicitly in the
\code{\link{comp_boot_emp}} function. So although passing
in \code{list(B = -15, m = -20)} will pass this function without errors,
these will be addressed explicitly in \code{\link{comp_boot_emp}} as
invalid inputs.}

\item{boot_sub}{(list) TODO: ADD}

\item{boot_res}{(list) : In the case of residual bootstrap the expected
input is of the form \code{list(B = 10)}. Note that technically \code{B}
should be a positive integer, but this assertion checking is handled
explicitly in the \code{\link{comp_boot_res}} function. So although passing
in \code{list(B = -15)} will pass this function without errors,
these will be addressed explicitly in \code{\link{comp_boot_res}} as
invalid inputs.}

\item{boot_mul}{(list) : In the case of multiplier bootstrap the expected
input is of the form \code{list(B = 10, weights_type = "rademacher")}.
Here the named element \code{weights_type} is optional
e.g. \code{list(B = 10)} is valid, or passed in as an explicit \code{NULL}
e.g. \code{list(B = 10, weights_type = NULL)}.
Note that technically \code{B} should be a positive integer, and
\code{weights_type} should be a character vector
(see \code{\link{comp_boot_mul}}), but this assertion checking is handled
explicitly in the \code{\link{comp_boot_mul}} function. So although passing
in \code{list(B = -15, m = "test")} will pass this function without errors,
these will be addressed explicitly in \code{\link{comp_boot_mul}} as
invalid inputs.}
}
\value{
A list containing the several types of variance estimates requested
by the  user, including the sandwich and the the variance returned by
\code{\link[stats]{lm}}.
}
\description{
\code{comp_mms_var} returns a list containing the requested estimates of the
variance, together with the assumptions behind which these estimates are
consistent.
}
\examples{
\dontrun{
# Simulate data from a linear model
set.seed(35542)
n <- 1e2
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)

# Fit the linear model using OLS (ordinary least squares)
mod_fit <- stats::lm(y ~ X)

# Run the multiplier bootstrap on the fitted (OLS) linear model
set.seed(162632)
out <- comp_mms_var(mod_fit,
boot_mul = list(B = 100, weights_type = "rademacher"),
boot_sub = list(B = 100, m = 50))

# print output
print(out)
}
}
\keyword{internal}
