% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-var.R
\name{comp_sandwich_qr_var}
\alias{comp_sandwich_qr_var}
\title{Compute the White sandwich estimator for ordinary least squares (OLS) regression.}
\usage{
comp_sandwich_qr_var(lm_object)
}
\arguments{
\item{lm_object}{(lm) : lm object}
}
\value{
(matrix) : White sandwich estimator of variance for OLS regression
}
\description{
Compute the White sandwich estimator for ordinary least squares (OLS) regression.
}
