% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-var.R
\name{comp_sandwich_var}
\alias{comp_sandwich_var}
\title{Compute the White sandwich estimator for ordinary least squares (OLS) regression.}
\usage{
comp_sandwich_var(X, residuals)
}
\arguments{
\item{X}{(matrix) : independent variables used in OLS fit}

\item{residuals}{(numeric) : residuals from OLS fit}
}
\value{
(matrix) : White sandwich estimator of variance for OLS regression
}
\description{
Compute the White sandwich estimator for ordinary least squares (OLS) regression.
}
