% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-var.R
\name{comp_var}
\alias{comp_var}
\title{Create an empirical or multiplier bootstrap summary table}
\usage{
comp_var(mod_fit, boot_emp = NULL, boot_res = NULL, boot_mul = NULL)
}
\arguments{
\item{mod_fit}{An lm (OLS) object}

\item{boot_emp}{to add}

\item{boot_res}{to add}

\item{boot_mul}{to add}
}
\value{
A summary statistics tibble for the bootstrap input.
}
\description{
\code{comp_var} creates an OLS bootstrap summary table for
empirical bootstrap output or multiplier bootstrap output as generated from
\code{\link{comp_boot_emp}} or
\code{\link{comp_boot_mul}}, respectively.
}
\examples{
\dontrun{
# Simulate data from a linear model
set.seed(35542)
n <- 1e2
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)

# Fit the linear model using OLS (ordinary least squares)
mod_fit <- stats::lm(y ~ X)

# Run the multiplier bootstrap on the fitted (OLS) linear model
set.seed(162632)
coef_emp_boot <- comp_boot_emp(mod_fit = mod_fit, B = 1000)
out <- comp_var(mod_fit, boot_mul = list(B = 100, weights_type = "rademacher"))

# print output
print(out)
}
}
