% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical.R
\name{diag_boot_emp_qqn}
\alias{diag_boot_emp_qqn}
\title{Normal QQ plot of the terms in an output of the bootstrap function}
\usage{
diag_boot_emp_qqn(boot_out)
}
\arguments{
\item{boot_out}{A tibble of the model's coefficients estimated (\code{term}
and \code{estimate}) on the bootstrapped datasets, the size of each
bootstrapped dataset (\code{m}), the size of the original dataset
(\code{n}), and the number of the bootstrap repetition (\code{b}).}
}
\value{
A ggplot2 object containing normal QQ plot for each regressor in
\code{boot_out}. Each panel corresponds to a different coefficient,
whose name appears in the panel's titles.
}
\description{
\code{diag_boot_emp_qqn} produces a normal QQPlot for each regressors included
in the estimates on the bootstrapped datasets. This function is a wrapper
for the \code{\link[ggplot2]{stat_qq}} function.
}
\examples{
\dontrun{
# Obtain normal QQ plot of the
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2 * 0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))

# Fit a linear model (OLS) to the data
mod_fit <- stats::lm(y ~ X1 + X2, df)
boot_out <- comp_boot_emp(mod_fit)

# Display the output
diag_boot_emp_qqn(boot_out)
}
}
