% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{focal_rwgt_var}
\alias{focal_rwgt_var}
\title{Obtain the "focal reweighting variable" model diagnostic.}
\usage{
focal_rwgt_var(mod_fit, coef_rwgt, term_chosen)
}
\arguments{
\item{mod_fit}{An object of class \code{lm} or \code{glm} to fit on the data. This object
should contain the formula, the data, and, in case of \code{glm}, the family.}

\item{coef_rwgt}{A tibble containing the number of the bootstrapped data set (\code{b}),
the size of each bootstrapped data set (\code{m}),
the value of the reweighting centers (\code{term_rwgt_center}) and of the reweighted term (\code{term_rwgt}),
and the estimates of the regression coefficients (\code{term} and \code{estimate}).}

\item{term_chosen}{Character corresponding to the coefficient to be analysed.}
}
\value{
A ggplot2 object which shows how the coefficients estimates
vary under reweighting of the regressor \code{term_chosen}.
The vertical axes represent the estimates of the coefficient of the regressors,
whose names appear in the panels titles, under reweighting of \code{term_chosen}.
Horizontal axes shows the values of the regressors.
Grey lines correspond to the traces of bootstrapped estimates forming the "spaghetti plot".
The black vertical lines indicate 95\% confidence intervals computed via the percentile method
for the estimates on the
bootstrapped data sets for each of the centers of reweighting (\code{term_rwgt_center}).
The black line in the middle corresponds to the mean of the estimates (which is
approximately equal to the estimates on the original reweighted data).
The blue dashed lines correspond exactly to the original estimate of the coefficients from \code{mod_fit}.
}
\description{
Obtain the "focal reweighting variable"  model diagnostic as described in
\insertCite{@see @buja2019modelsasapproximationspart2;textual}{maar}.
}
\details{
The function shows how the estimates of all regressors vary under reweighting
of one regressor, specified in \code{term_chosen}, and compare with the original
estimates from \code{mod_fit}.
}
\examples{
\dontrun{
# Get focal reweighting variable plot of X1
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2 * 0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))
mod_fit <- stats::lm(y ~ X1 + X2, df)
ols_rwgt <- comp_coef_rwgt(mod_fit, c("X1", "X2"), B = 300)

# Display the output
focal_rwgt_var(mod_fit, ols_rwgt, "X1")
}

}
\references{
\insertAllCited{}
}
