% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{focal_slope}
\alias{focal_slope}
\title{Obtain the "focal slope" model diagnostic}
\usage{
focal_slope(coef_rwgt, term_chosen)
}
\arguments{
\item{coef_rwgt}{A tibble containing the number of the bootstrapped data set (b),
the names of the regressors under reweighting, and the sets of model estimates}

\item{term_chosen}{Character corresponding to the coefficient to be analysed}
}
\value{
A ggplot2 object which shows how the coefficient of one regressor of interest (term_chosen)
varies under reweighting of the regressors.
Vertical axis = regression coefficient of PercVacant (in all frames);
horizontal axes and panels titles = regressors.
The vertical axis corresponds the size of the coefficient of
the regressor of interest.
Grey lines correspond to the traces of bootstrapped estimates forming the "spaghetti plot".
The black vertical lines indicate 95\% confidence intervals for the estimates on the
bootstrapped data sets for each of the centers of reweighting (term_rwgt_center).
The black horizontal line corresponds to the medians of the estimates.
The text in the top left corner of each panel shows the results
of a t-test for testing the statistical significance of the difference of the means
of the distributions of boostrap estimates between the two most extreme centers.
}
\description{
Obtain the "focal slope" model diagnostic as described in
\insertCite{@see @buja2019modelsasapproximationspart2;textual}{maar}.
The diagnostic is useful to investigate interaction terms that have not been included
in the model in the first place.
}
\examples{
\dontrun{
# Get focal slope of X1
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2*0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))
mod_fit <- stats::lm(y ~ X1+X2, df)
ols_rwgt <- comp_coef_rwgt(mod_fit, c('X1', 'X2'), B=300)

# Display the output
focal_slope(ols_rwgt, 'X1')
}
}
\references{
\insertAllCited{}
}
