% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-multiplier-boot-var.R
\name{gen_multiplier_bootstrap_weights}
\alias{gen_multiplier_bootstrap_weights}
\title{Generate Different Types of Multiplier Bootstrap Weights}
\usage{
gen_multiplier_bootstrap_weights(n, weights_type)
}
\arguments{
\item{n}{The number of random Multiplier bootstrap weights to generate}

\item{weights_type}{The type of Multiplier bootstrap weights to generate.
Based on the \code{weighttype} option in the \code{Stata boottest} package,
this can only take the following five prespecified values
\code{"rademacher", "mammen", "webb", "std_gaussian", "gamma"}. A brief
description of each type is noted as below. The \code{"rademacher"} weights
are sampled from the two point Rademacher distribution which has takes values
\eqn{\{1, 1\}}, with equal probability. The \code{"mammen"} weights
are sampled from the two point Mammen distribution which has takes values
\eqn{\{\phi, 1 - \phi\}}, with probabilities
\eqn{\{\frac{\phi}{\sqrt{5}}, 1 - \frac{\phi}{\sqrt{5}}\}}, respectively. The
\code{"webb"} weights are sampled from the six point Webb distribution
which has takes values \eqn{\{\pm \sqrt{\frac{3}{2}}, \pm \sqrt{\frac{1}{2}}, \pm 1 \}},
with equal probability. The \code{"std_gaussian"} weights are sampled from
the standard Gaussian (normal) distribution with mean = \eqn{0},
and variance = \eqn{1}. Finally, the \code{"gamma"} weights are sampled from
the Gamma distribution with shape parameter = \eqn{4}, and scale parameter = \eqn{1}.
For more details on these weight types see \insertCite{@see @roodman2019fastwildinferencestataboottest;textual}{maar} for more details.}
}
\value{
A numeric vector of n (sampled with replacement) random Multiplier
bootstrap weights based on the specified Multiplier weights type
}
\description{
A Helper function for \code{\link[maar]{comp_multiplier_bootstrap_var}}
to generate different types of Multiplier Bootstrap
Weights. This section is inspired by the \code{weighttype} option in the
\code{Stata boottest} package.
}
\examples{
\dontrun{
# Number of Multiplier weights to generate
n <- 1000

# Generate the different type of Multiplier Weights
rademacher_w <- gen_multiplier_bootstrap_weights(n = n,
                                                 weights_type = "rademacher")
mammen_w <- gen_multiplier_bootstrap_weights(n = n, weights_type = "mammen")
webb_w <- gen_multiplier_bootstrap_weights(n = n, weights_type = "webb")
std_gaussian_w <- gen_multiplier_bootstrap_weights(n = n,
                                                   weights_type = "std_gaussian")
gamma_w <- gen_multiplier_bootstrap_weights(n = n, weights_type = "gamma")
}
}
\references{
\insertAllCited
}
