% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-confint.R
\name{get_confint2}
\alias{get_confint2}
\title{Get the tidy variance summary and confidence intervals for the coefficients of
an object of class "maars_lm"}
\usage{
get_confint2(
  mod_fit,
  level = 0.95,
  sand = TRUE,
  boot_emp = FALSE,
  boot_mul = FALSE,
  boot_res = FALSE,
  well_specified = FALSE,
  tidy = TRUE
)
}
\arguments{
\item{mod_fit}{(maars_lm, lm) A fitted OLS \code{maars_lm, lm} class object}

\item{level}{(double) : numeric value between 0 and 1 indicating the
confidence level (e.g., 0.95)}

\item{sand}{(logical) : \code{TRUE} if sandwich estimator output is required,
\code{FALSE} to exclude this output from the request}

\item{boot_emp}{(logical) : \code{TRUE} if empirical bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{boot_mul}{(logical) : \code{TRUE} if multiplier bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{boot_res}{(logical) : \code{TRUE} if residual bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{well_specified}{(logical) : \code{TRUE} if lm standard errors
(well specified) output is required, \code{FALSE} to exclude this output
from the request}

\item{tidy}{(logical) : \code{TRUE} if user requires a
\code{\link[broom]{tidy}} style variance summary, \code{FALSE} if the user
requires the compressed variance summary joined by the common
"term", "estimate" values for the OLS fit. In this output each variance
type has the required standard error term as a separate column with the
variance type abbreviation as the suffix. That is, for sandwich estimator
the 'p.value' column is renamed to 'p.value.sand'.}
}
\value{
(tibble) : Combined standard error and confidence intervals
summary from a fitted OLS \code{maars_lm, lm} class object
}
\description{
Get the tidy variance summary and confidence intervals from a fitted OLS \code{maars_lm, lm}
class object
}
\details{
The function generates the same output as \code{\link{get_boot_summary}},
but also includes lower and upper confidence intervals for the regression
coefficients according to the significance level specified in the "level"
argument.
}
\examples{
\dontrun{
# TODO: Add here
}
}
