% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{get_mms_assumptions_cli}
\alias{get_mms_assumptions_cli}
\title{Get assumptions for a \code{maars_lm, lm} object, for a specific
variance type}
\usage{
get_mms_assumptions_cli(title, assumptions)
}
\arguments{
\item{title}{TODO: Add later}

\item{assumptions}{TODO: Add later}
}
\value{
TODO: Add later
}
\description{
Get assumptions for a \code{maars_lm, lm} object, for a specific
variance type
}
\examples{
\dontrun{
# TODO: Add later
}
}
\keyword{internal}
