% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-var.R
\name{get_mms_comp_var_ind}
\alias{get_mms_comp_var_ind}
\title{Generates list containing estimates for an individual variance type}
\usage{
get_mms_comp_var_ind(
  var_type_abb,
  summary_tbl = NULL,
  cov_mat = NULL,
  B = NULL,
  m = NULL,
  n = NULL,
  weights_type = NULL,
  boot_out = NULL
)
}
\arguments{
\item{var_type_abb}{(character) : The abbreviation used to classify the type
of variance. The choice is one of the following
\code{"lm", "sand", "emp", "sub", "res", "mul"}.}

\item{summary_tbl}{(tibble) : The summary table of standard errors for the
type of variance.}

\item{cov_mat}{(matrix) : The covariance matrix for the type of variance.}

\item{B}{(integer) : The number of bootstrap replications.}

\item{m}{(integer) : The number of observations to be sampled with or without
replacement from the dataset for each bootstrap repetition.}

\item{n}{(integer) : The number of observations from the original modeling
dataset.}

\item{weights_type}{: The type of multiplier bootstrap weights to generate.
Based on the \code{weighttype} option in the Stata \code{boottest} package,
this can only take the following five prespecified values
\code{"rademacher", "mammen", "webb", "std_gaussian", "gamma"}.
For more details see the documentation for
\code{\link{comp_boot_mul_wgt}}. The default value is "rademacher".}

\item{boot_out}{(tibble) : A tibble of the boostrap replication datasets
used in \code{\link{comp_boot_mul}}, \code{\link{comp_boot_res}}. In all
other variance types this should be set to \code{NULL}.}
}
\value{
(list) : A list containing essential computed variance output for the
type of variance specified. The computed parameter types are the common
output for the following functions: \code{\link{comp_lm_var}},
\code{\link{comp_sand_var}}, \code{\link{comp_boot_emp}},
\code{\link{comp_boot_mul}}, \code{\link{comp_boot_res}}.
}
\description{
\code{get_mms_comp_var_ind} generates the final output list for an
individual variance type. This is a helper function to be used in the
following functions: \code{\link{comp_lm_var}}, \code{\link{comp_sand_var}},
\code{\link{comp_boot_emp}}, \code{\link{comp_boot_mul}},
\code{\link{comp_boot_res}}.
}
\keyword{internal}
