% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{get_mms_print_cli}
\alias{get_mms_print_cli}
\title{Print method display for a \code{maars_lm, lm} object}
\usage{
get_mms_print_cli(mod_fit, all_emoji_titles, all_assumptions)
}
\arguments{
\item{mod_fit}{TODO: Add later}

\item{all_emoji_titles}{TODO: Add later}

\item{all_assumptions}{TODO: Add later}
}
\value{
TODO: Add later
}
\description{
Print method display for a \code{maars_lm, lm} object
}
\examples{
\dontrun{
# TODO: Add later
}
}
\keyword{internal}
