% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-summary.R
\name{get_mms_rnm_cols_suff}
\alias{get_mms_rnm_cols_suff}
\title{Adds a variance type abbreviation suffix to variance summary tibbles}
\usage{
get_mms_rnm_cols_suff(var_summary, var_type_abb)
}
\arguments{
\item{var_summary}{TODO: Add later}

\item{var_type_abb}{TODO: Add later}
}
\value{
(\code{tibble}) : TODO Add later
}
\description{
Adds a variance type abbreviation suffix to variance summary tibbles
}
\examples{
\dontrun{
# TODO: Add here
}
}
\keyword{internal}
