% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{get_mms_summary_joined_cli}
\alias{get_mms_summary_joined_cli}
\title{Print interleaved summary for a \code{maars_lm, lm} object, for all
variance types}
\usage{
get_mms_summary_joined_cli(
  all_emoji_titles,
  all_assumptions,
  summary_joined,
  digits = 3
)
}
\arguments{
\item{all_emoji_titles}{TODO: Add later}

\item{all_assumptions}{TODO: Add later}

\item{summary_joined}{TODO: Add later}

\item{digits}{TODO: Add later}
}
\value{
TODO: Add later
}
\description{
Print interleaved summary for a \code{maars_lm, lm} object, for all
variance types
}
\examples{
\dontrun{
# TODO: Add later
}
}
\keyword{internal}
