% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{get_mms_summary_print_lm_style}
\alias{get_mms_summary_print_lm_style}
\title{Print summary in the style of \code{lm}}
\usage{
get_mms_summary_print_lm_style(var_summary, digits)
}
\arguments{
\item{var_summary}{TODO: Add later}

\item{digits}{TODO: Add later}
}
\value{
TODO: Add later
}
\description{
Print summary in the style of \code{lm}
}
\examples{
\dontrun{
# TODO: Add later
}
}
\keyword{internal}
