% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{get_mms_summary_split_cli}
\alias{get_mms_summary_split_cli}
\title{Print interleaved summary for a \code{maars_lm, lm} object, for a
specific variance type}
\usage{
get_mms_summary_split_cli(title, summary, assumptions, digits = 3)
}
\arguments{
\item{title}{TODO: Add later}

\item{summary}{TODO: Add later}

\item{assumptions}{TODO: Add later}

\item{digits}{TODO: Add later}
}
\value{
TODO: Add later
}
\description{
Print interleaved summary for a \code{maars_lm, lm} object, for a
specific variance type
}
\examples{
\dontrun{
# TODO: Add later
}
}
\keyword{internal}
