% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-diag-plot.R
\name{get_ols_diag_plots}
\alias{get_ols_diag_plots}
\title{Get a list of standard diagnostic plots for \code{\link[stats]{lm}},
in the \code{link[ggplot2]} package style.}
\usage{
get_ols_diag_plots(mod_fit)
}
\arguments{
\item{mod_fit}{(maars_lm, lm) A fitted OLS \code{maars_lm, lm} class object}
}
\value{
A list of standard diagnostic plots for OLS in the
\code{link[ggplot2]} package style.
}
\description{
Get a list of standard diagnostic plots for \code{\link[stats]{lm}},
in the \code{link[ggplot2]} package style.
}
\examples{
\dontrun{
set.seed(1243434)

# generate data
n <- 1e3
X_1 <- stats::rnorm(n, 0, 1)
X_2 <- stats::rnorm(n, 10, 20)
eps <- stats::rnorm(n, 0, 1)

# OLS data and model
y <- 2 + X_1 * 1 + X_2 * 5 + eps
lm_fit <- stats::lm(y ~ X_1 + X_2)
mms_fit <- lm_fit \%>\% comp_var(
  mod_fit = .,
  boot_emp = list(B = 100, m = 50),
  boot_res = list(B = 50)
)

# Produce and sequentially display diagnostic plots
plot(mms_fit)
}
}
\keyword{internal}
