% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-var.R
\name{get_summary}
\alias{get_summary}
\title{get_summary computes the statistics based on boot_out and boot_type}
\usage{
get_summary(mod_fit, boot_out, boot_type)
}
\arguments{
\item{mod_fit}{An "lm" (OLS) object.}

\item{boot_out}{A tibble of the model's coefficients estimated (\code{term}
and \code{estimate}) on the bootstrapped datasets, the size of each
bootstrapped dataset (\code{m}), the size of the original dataset
(\code{n}), and the number of the bootstrap repetition (\code{b}).}

\item{boot_type}{A character specifying the bootstrap type. It can be be "emp"
for output from \code{\link{comp_boot_emp}}, "mul"
for output from \code{\link{comp_boot_mul}}, or
"res" for output from \code{\link{comp_boot_res}}.}
}
\value{
A tibble containing the summary statistics for the model: terms,
coefficients estimates, t-statistics, and p-values. These statistics are
based on the output of the bootstrap passed in \code{boot_out}.


}
\description{
\code{get_summary} returns a tibble containing the model's statistics
based on the coefficients estimates (\code{boot_out}) obtained via
\code{boot_type} bootstrap (e.g., empirical).
}
\keyword{internal}
