% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-summary.R
\name{get_var_tidy_summary_ind}
\alias{get_var_tidy_summary_ind}
\title{An individual variance table summary function that adds the variance type
suffix to the table columns}
\usage{
get_var_tidy_summary_ind(comp_mms_var_dat)
}
\arguments{
\item{comp_mms_var_dat}{TODO Add}
}
\value{
TODO Add
}
\description{
An individual variance table summary function that adds the variance type
suffix to the table columns
}
\keyword{internal}
