% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-package.R
\docType{package}
\name{maars-package}
\alias{maars}
\alias{maars-package}
\title{maars: Tidy Inference Under Misspecified Statistical Models in R}
\description{
To learn more about maars, start with the vignettes:
\code{browseVignettes(package = "maars")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shamindras/maars}
  \item \url{https://shamindras.github.io/maars/}
  \item Report bugs at \url{https://github.com/shamindras/maars/issues}
}

}
\author{
\strong{Maintainer}: Shamindra Shrotriya \email{shamindra.shrotriya@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Riccardo Fogliato \email{rfogliat@andrew.cmu.edu} [copyright holder]
  \item Arun Kumar Kuchibhotla \email{karun3kumar@gmail.com} [copyright holder]
}

}
\keyword{internal}
