% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-boot-var.R
\name{multiplier_bootstrap}
\alias{multiplier_bootstrap}
\title{A wrapper to replicate the multiplier bootstrap calculation for an OLS fitted dataset}
\usage{
multiplier_bootstrap(lm_fit, B = 100)
}
\arguments{
\item{lm_fit}{An lm (OLS) object}

\item{B}{The number of bootstrap replications}
}
\value{
A tibble of the bootstrap calculations
}
\description{
The multiplier bootstrap calculated for a single replication
for an OLS fitted dataset is given by the following expression:
\deqn{\frac{1}{n}\sum_{i=1}^{n} e_{i}\widehat{J}^{-1}X_{i}(Y_{i}-X_{i}^{T} \widehat{\beta})}
This wrapper function calls on the helper function \code{\link{multiplier_single_bootstrap}}. See it's documentation
for how a single instance is run.
}
\examples{
\dontrun{
# Simulate data from a linear model
set.seed(35542)
n <- 1e2
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)

# Fit the linear model using OLS (ordinary least squares)
lm_fit <- stats::lm(y ~ X)

# Run the multiplier bootstrap on the fitted (OLS) linear model
set.seed(162632)
multiplier_bootstrap(lm_fit, B = 15)
}
}
