% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maa-ols.R
\name{my_sum}
\alias{my_sum}
\title{Returns the sum of 2 numbers (or numeric vectors)}
\usage{
my_sum(a, b)
}
\arguments{
\item{a}{(numeric) : A numeric vector (can contain just a single number)}

\item{b}{(numeric) : A numeric vector (can contain just a single number)}
}
\value{

}
\description{
Returns the sum of 2 numbers (or numeric vectors)
}
