% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{plot.maars_lm}
\alias{plot.maars_lm}
\title{Plot \code{maars_lm, lm} object}
\usage{
\method{plot}{maars_lm}(x, which = NULL, ...)
}
\arguments{
\item{x}{(\code{maars_lm, lm}) : A fitted \code{maars_lm, lm} OLS object}

\item{which}{(numeric vector) : if a subset of the plots is required, specify a
subset of the numbers 1:6.}

\item{...}{Additional arguments passed to methods.}
}
\value{
TODO: Describe plots
}
\description{
Plot \code{maars_lm, lm} object
}
\examples{
\dontrun{
set.seed(454354534)

# generate data ----
n <- 1e3
X_1 <- stats::rnorm(n, 0, 1)
X_2 <- stats::rnorm(n, 10, 20)
eps <- stats::rnorm(n, 0, 1)

# Let's generate data and fit a well-specified OLS data and model ----
y <- 2 + X_1 * 1 + X_2 * 5 + eps
lm_fit <- stats::lm(y ~ X_1 + X_2)
summary(lm_fit)

# Fit our first maars_lm object i.e. comp_var1
comp_var1 <- comp_var(mod_fit = lm_fit,
                      boot_emp = list(B = 50, m = 200),
                      boot_res = NULL,
                      boot_mul = list(B = 60))

# Plot our maars_lm object
plot(comp_var1)
}
}
