% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-summary.R
\name{print.maars_lm}
\alias{print.maars_lm}
\title{Print \code{maars_lm} object}
\usage{
\method{print}{maars_lm}(x, ...)
}
\arguments{
\item{x}{A \code{maars_lm} object.}

\item{...}{Additional arguments passed to \code{print.lm()} to print the
object.}
}
\description{
Calls \code{print} on a \code{\link[stats]{lm}} object.
}
