% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{print.summary.maars_lm}
\alias{print.summary.maars_lm}
\title{Print summary of \code{maars_lm} object}
\usage{
\method{print}{summary.maars_lm}(x, ...)
}
\arguments{
\item{x}{A \code{maars_lm} object.}

\item{...}{Additional arguments}
}
\description{
Calls \code{print.summary.maars_lm} on a \code{summary} of a \code{maars_lm} object.
}
