% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{set_ggplot2_theme}
\alias{set_ggplot2_theme}
\title{Set the custom \code{\link[ggplot2]{theme}} for our plots}
\usage{
set_ggplot2_theme(
  ggplot_obj,
  axis_text_size = 17,
  axis_title_size = 17,
  strip_text_x_size = 17,
  legend_text_size = 17,
  legend_title_size = 17,
  legend_position = "bottom"
)
}
\arguments{
\item{ggplot_obj}{A \code{\link[ggplot2]{ggplot2}} object}

\item{axis_text_size}{A positive integer to define the the size of the x and
y axis text in our \code{\link[ggplot2]{ggplot2}} object}

\item{axis_title_size}{A positive integer to define the the size of the x and
y axis text in our \code{\link[ggplot2]{ggplot2}} object}

\item{strip_text_x_size}{A positive integer to define the the size of the x and
y axis text in our \code{\link[ggplot2]{ggplot2}} object}

\item{legend_text_size}{A positive integer to define the the size of the x and
y axis text in our \code{\link[ggplot2]{ggplot2}} object}

\item{legend_title_size}{A positive integer to define the the size of the x and
y axis text in our \code{\link[ggplot2]{ggplot2}} object}

\item{legend_position}{A character value the position of legends
("none", "left", "right", "bottom", "top", or two-element numeric vector)}
}
\value{

}
\description{
\code{set_ggplot2_theme} is a wrapper to set a standard custom
\code{\link[ggplot2]{theme}} for all of our output plots. It contains
only the necessary parameters we require to change in the
\code{\link[ggplot2]{theme}} function, and takes the remaining arguments
as the default values.
}
\examples{
\dontrun{
library(ggplot2)
p1 <- ggplot2::ggplot(mtcars, aes(x = wt, y = mpg, size = disp)) +
    ggplot2::geom_point()

# Set the theme for a ggplot2 object with the default options
out <- p1 \%>\% maars::set_ggplot2_theme(ggplot_obj = .)
}
}
