% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maars-lm.R
\name{summary.maars_lm}
\alias{summary.maars_lm}
\title{Summary of \code{maars_lm} object}
\usage{
\method{summary}{maars_lm}(
  object,
  sand = FALSE,
  boot_emp = FALSE,
  boot_res = FALSE,
  boot_mul = FALSE,
  well_specified = FALSE,
  digits = 3,
  ...
)
}
\arguments{
\item{object}{A fitted "maars_lm" object.}

\item{sand}{(logical) : \code{TRUE} if sandwich estimator output is required,
\code{FALSE} to exclude this output from the request.}

\item{boot_emp}{(logical) : \code{TRUE} if empirical bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request.}

\item{boot_res}{(logical) : \code{TRUE} if residual bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request.}

\item{boot_mul}{(logical) : \code{TRUE} if multiplier bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{well_specified}{(logical) : \code{TRUE} if lm standard errors.
(well specified) output is required, \code{FALSE} to exclude this output
from the request.}

\item{digits}{(integer) : Rounding digits used in some of the function's output.}

\item{...}{Additional arguments.}
}
\description{
Summary method for class "maars_lm".
}
