% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-boot-multiplier.R
\name{comp_boot_mul}
\alias{comp_boot_mul}
\title{A wrapper to replicate the multiplier bootstrap calculation for an OLS fitted dataset}
\usage{
comp_boot_mul(mod_fit, B, weights_type)
}
\arguments{
\item{mod_fit}{An \code{\link[stats]{lm}} (OLS) object.}

\item{B}{The number of bootstrap replications.}

\item{weights_type}{The type of multiplier bootstrap weights to generate.
Based on the \code{weighttype} option in the Stata \code{boottest} package,
this can only take the following five prespecified values
\code{"rademacher", "mammen", "webb", "std_gaussian", "gamma"}.
For more details see the documentation for
\code{\link{comp_boot_mul_wgt}}.}
}
\value{
A tibble of the bootstrap calculations.
}
\description{
The multiplier bootstrap calculated for a single replication
for an OLS fitted dataset is given by the following expression:
\deqn{\frac{1}{n}\sum_{i=1}^{n} e_{i}\widehat{J}^{-1}X_{i}(Y_{i}-X_{i}^{T} \widehat{\beta})}
This wrapper function calls on the helper function \code{\link{comp_boot_mul_ind}}. See its documentation
for how a single instance is run.
}
\examples{
\dontrun{
# Simulate data from a linear model
set.seed(35542)
n <- 1e2
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)

# Fit the linear model using OLS (ordinary least squares)
lm_fit <- stats::lm(y ~ X)

# Run the multiplier bootstrap on the fitted (OLS) linear model
set.seed(162632)
comp_boot_mul(lm_fit, B = 15, weights_type = 'std_gaussian')
}
}
\keyword{internal}
