% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{comp_coef_rwgt}
\alias{comp_coef_rwgt}
\title{Obtain estimates of a GLM or OLS estimates under reweighting of one or multiple regressors}
\usage{
comp_coef_rwgt(
  mod_fit,
  terms_to_rwgt,
  B = 100,
  m = NULL,
  grid_bins_centers = NULL,
  grid_method = "quantiles",
  n_grid = 9
)
}
\arguments{
\item{mod_fit}{An object of class "lm" or "glm" to fit on the data. This object
should contain the formula, the data, and, in case of "glm", the family}

\item{terms_to_rwgt}{A character corresponding to the regressor to be reweighted}

\item{boot_samples}{A list of data sets that have been bootstrapped from the
original data set. Each data set needs to contain a column "n_obs" which indicates
the order of each observation in the original data}

\item{term_to_rwgt_centers}{A vector of numeric values corresponding to the
centers for the reweighting procedure}
}
\value{
A tibble containing the number of the bootstrapped data set (b),
the size of each bootstrapped data set (m),
the names of the regressors under reweighting,
and the sets of model estimates
}
\description{
Obtain estimates of a GLM or OLS estimates under reweighting of one or multiple regressor.
The function returns a set of estimates for each regressor that is reweighted,
for each bootstrapped data set, and for each center of reweighting.
}
\examples{
\dontrun{
# Get OLS estimates under reweighting of all regressors
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2*0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))
mod_fit <- stats::lm(y ~ X1+X2, df)
ols_rwgt <- comp_coef_rwgt(mod_fit, c('X1', 'X2'))

# Display the output
print(ols_rwgt)
}
}
