% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical-var.R
\name{comp_cond_model}
\alias{comp_cond_model}
\title{Fit OLS or GLM on the data}
\usage{
comp_cond_model(mod_fit, data, weights = NULL)
}
\arguments{
\item{mod_fit}{An object of class "lm" or "glm" to fit on the data. This object
should contain the formula, the data, and, in case of "glm", the family}

\item{data}{A tibble or data frame containing the data set on which the model
will be trained}

\item{weights}{A character corresponding to the name of the weights
feature name in the data}
}
\value{
A tibble containing the estimated coefficients of the model
}
\description{
Fit OLS or GLM on the data
}
\examples{
\dontrun{
# Estimate OLS from a bootstraped data set
n <- 1e3
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)
lm_fit <- stats::lm(y ~ X)
boot <- comp_empirical_bootstrap_samples(model.frame(lm_fit))$data[[1]]
mod <- comp_cond_model(lm_fit, boot)

# Display the output
print(mod)

#' # Estimate OLS from a data set with weights
n <- 1e3
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 1)
mod_fit <- stats::lm(y ~ X)
df <- tibble::tibble(y = y, X = X, weights = 1:length(X))
mod <- comp_cond_model(mod_fit, df, 'weights')

# Display the output
print(mod)
}
}
