% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{focal_rwgt_var}
\alias{focal_rwgt_var}
\title{Obtain the "focal reweighting variable" model diagnostic}
\usage{
focal_rwgt_var(mod_fit, coef_rwgt, term_chosen)
}
\arguments{
\item{mod_fit}{An object of class "lm" or "glm" to fit on the data. This object
should contain the formula, the data, and, in case of "glm", the family}

\item{coef_rwgt}{A tibble containing the number of the bootstrapped data set (b),
the names of the regressors under reweighting, and the sets of model estimates}

\item{term_chosen}{Character corresponding to the coefficient to be analysed}
}
\value{
A ggplot2 object which shows how the coefficients estimates (vertical axes)
of the different regressors (title of each panel) vary under
reweighting of one regressor (horizontal axis).
Vertical axes and panels titles = estimates of the regression coefficients and
names of the corresponding regressors.
Horizontal axes = value of the regressor of interest.
Grey lines correspond to the traces of bootstrapped estimates forming the "spaghetti plot".
The black vertical lines indicate 95\% confidence intervals for the estimates on the
bootstrapped data sets for each of the centers of reweighting (term_rwgt_center).
The black line in the middle corresponds to the mean of the estimates (which is
approximately equal to the OLS estimates on the original data).
}
\description{
Obtain the "focal reweighting variable" model diagnostic described in
\insertCite{@see @buja2019modelsasapproximationspart2;textual}{maar}.
}
\examples{
\dontrun{
# Get focal reweighting variable plot of X1
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2 * 0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))
mod_fit <- stats::lm(y ~ X1 + X2, df)
ols_rwgt <- comp_coef_rwgt(mod_fit, c("X1", "X2"), B = 300)

# Display the output
focal_rwgt_var(mod_fit, ols_rwgt, "X1")
}

}
\references{
\insertAllCited{}
}
