% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{focal_slope}
\alias{focal_slope}
\title{Obtain the "focal slope" model diagnostic}
\usage{
focal_slope(mod_fit, coef_rwgt, term_chosen)
}
\arguments{
\item{mod_fit}{An object of class "lm" or "glm" to fit on the data. This object
should contain the formula, the data, and, in case of "glm", the family}

\item{coef_rwgt}{A tibble containing the number of the bootstrapped data set (b),
the names of the regressors under reweighting, and the sets of model estimates}

\item{term_chosen}{Character corresponding to the coefficient to be analysed}
}
\value{
A ggplot2 object which shows how the coefficient of one regressor of interest (term_chosen)
varies under reweighting of the regressors.
Vertical axes = estimates of the regression coefficient of interest.
Horizontal axes and panels titles = values and names of the regressors respectively.
Grey lines correspond to the traces of bootstrapped estimates forming the "spaghetti plot".
The black vertical lines indicate 95\% confidence intervals for the estimates on the
bootstrapped data sets for each of the centers of reweighting (term_rwgt_center).
The black line in the middle corresponds to the mean of the estimates (which is
approximately equal to the OLS estimates on the original data).
}
\description{
Obtain the "focal slope" model diagnostic as described in
\insertCite{@see @buja2019modelsasapproximationspart2;textual}{maar}.
The diagnostic is useful to investigate interaction terms that have not been included
in the model in the first place.
}
\examples{
\dontrun{
# Get focal slope of X1
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2 * 0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))
mod_fit <- stats::lm(y ~ X1 + X2, df)
ols_rwgt <- comp_coef_rwgt(mod_fit, c("X1", "X2"), B = 300)

# Display the output
focal_slope(mod_fit, ols_rwgt, "X1")
}
}
\references{
\insertAllCited{}
}
