% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-sandwich-var.R
\name{get_summary}
\alias{get_summary}
\title{get_summary computes the statistics based on boot_out and boot_type}
\usage{
get_summary(mod_fit, boot_out, boot_type)
}
\arguments{
\item{boot_type}{}
}
\value{

}
\description{
get_summary computes the statistics based on boot_out and boot_type
}
\keyword{internal}
