% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-summary.R
\name{get_summary}
\alias{get_summary}
\title{Get the tidy variance summary from a fitted OLS \code{maars_lm, lm}
class object}
\usage{
get_summary(
  mod_fit,
  sand = NULL,
  boot_emp = NULL,
  boot_mul = NULL,
  boot_res = NULL,
  well_specified = NULL
)
}
\arguments{
\item{mod_fit}{(\code{maars_lm, lm}) A fitted OLS \code{maars_lm, lm} class
object}

\item{sand}{(logical) : \code{TRUE} if sandwich estimator output is required,
\code{FALSE} to exclude this output from the request}

\item{boot_emp}{(logical) : \code{TRUE} if empirical bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{boot_mul}{(logical) : \code{TRUE} if multiplier bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{boot_res}{(logical) : \code{TRUE} if residual bootstrap standard error
output is required, \code{FALSE} to exclude this output from the request}

\item{well_specified}{(logical) : \code{TRUE} if lm standard errors
(well specified) output is required, \code{FALSE} to exclude this output
from the request}
}
\value{
(tibble) : Combined standard error summary from a fitted
OLS \code{maars_lm, lm} class object
}
\description{
Get the tidy variance summary from a fitted OLS \code{maars_lm, lm}
class object
}
\examples{
\dontrun{
set.seed(1243434)

# generate data
n <- 1e3
X_1 <- stats::rnorm(n, 0, 1)
X_2 <- stats::rnorm(n, 10, 20)
eps <- stats::rnorm(n, 0, 1)

# OLS data and model
y <- 2 + X_1 * 1 + X_2 * 5 + eps
lm_fit <- stats::lm(y ~ X_1 + X_2)


# Empirical Bootstrap check
set.seed(454354534)
mms_var <- comp_var(
  mod_fit = lm_fit, boot_emp = list(B = 20, m = 200),
  boot_res = list(B = 30)
)

get_summary(mms_var)

# compute variance with multiplier bootstrap now
mms_var2 <- comp_var(
  mod_fit = lm_fit,
  boot_mul = list(B = 100)
)

get_summary(mms_var2)

}
}
