% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{nonlinearity_detection}
\alias{nonlinearity_detection}
\title{Obtain the "nonlinearity detection" model diagnostic}
\usage{
nonlinearity_detection(coef_rwgt)
}
\arguments{
\item{coef_rwgt}{A tibble containing the number of the bootstrapped data set (b),
the names of the regressors under reweighting, and the sets of model estimates}
}
\value{
A ggplot2 object which shows how the coefficients estimates
vary under reweighting of their own regressors.
Vertical axes = estimates of the regression coefficients.
Horizontal axes and panels titles = regressors.
Grey lines correspond to the traces of bootstrapped estimates forming the "spaghetti plot".
The black vertical lines indicate 95\% confidence intervals for the estimates on the
bootstrapped data sets for each of the centers of reweighting (term_rwgt_center).
The black line in the middle corresponds to the mean of the estimates (which is
approximately equal to the OLS estimates on the original data).
}
\description{
Obtain the "nonlinearity detection" model diagnostic described in
\insertCite{@see @buja2019modelsasapproximationspart2;textual}{maar}.
This tool provides insights into marginal nonlinear behavior of response surfaces.
}
\examples{
\dontrun{
# Get focal slope of X1
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2 * 0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))
mod_fit <- stats::lm(y ~ X1 + X2, df)
ols_rwgt <- comp_coef_rwgt(mod_fit, c("X1", "X2"), B = 300)

# Display the output
nonlinearity_detection(ols_rwgt)
}
}
\references{
\insertAllCited{}
}
