% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-empirical-var.R
\name{qqnorm_bootstrap}
\alias{qqnorm_bootstrap}
\title{Normal QQ plot of the terms in an output of the bootstrap function}
\usage{
qqnorm_bootstrap(boot_out)
}
\arguments{
\item{boot_out}{A tibble of the model's coefficients estimated (\code{term} and \code{estimate})
on the bootstrapped data sets, the size of each bootstrapped data set (\code{m}),
the size of the original data set (\code{n}),
and the number of the bootstrap repetition (\code{b}).}
}
\value{
A ggplot2 object containing normal QQ plot for each regressor in \code{boot_out}.
Each panel corresponds to a different coefficient, whose name appears in the panel's titles.
}
\description{
\code{qqnorm_bootstrap} produces a normal QQPlot for each regressors included in the
estimates on the bootstrapped data sets.
This function is a wrapper for the \code{ggplot2::stat_qq} function.
}
\examples{
\dontrun{
# Obtain normal QQ plot of the
n <- 1e3
X1 <- stats::rnorm(n, 0, 1)
X2 <- stats::rnorm(n, 0, 3)
y <- 2 + X1 + X2 * 0.3 + stats::rnorm(n, 0, 1)
df <- tibble::tibble(y = y, X1 = X1, X2 = X2, n_obs = 1:length(X1))
mod_fit <- stats::lm(y ~ X1 + X2, df)
boot_out <- comp_empirical_bootstrap(mod_fit)

# Display the output
qqnorm_bootstrap(boot_out)
}
}
